requirements.txt
alembic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
sqlalchemy
flask_migrate
pandas
flask_script
jsonify
flask
wget
oauth2client
flask_sqlalchemy
google-cloud-storage
requests
uwsgi
